Refactor processing
