+++ internal/processing/istio/authorization_policy_helpers.go
@@ -1,7 +1,8 @@
-package processing
+package istio
 
 import (
 	"fmt"
+
 	"istio.io/api/security/v1beta1"
 
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
@@ -9,20 +10,20 @@ import (
 	securityv1beta1 "istio.io/client-go/pkg/apis/security/v1beta1"
 )
 
-func modifyRequestAuthentication(existing, required *securityv1beta1.RequestAuthentication) {
+func modifyAuthorizationPolicy(existing, required *securityv1beta1.AuthorizationPolicy) {
 	existing.Spec = *required.Spec.DeepCopy()
 }
 
-func generateRequestAuthentication(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule, additionalLabels map[string]string) *securityv1beta1.RequestAuthentication {
+func generateAuthorizationPolicy(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule, additionalLabels map[string]string) *securityv1beta1.AuthorizationPolicy {
 	namePrefix := fmt.Sprintf("%s-", api.ObjectMeta.Name)
 	namespace := api.ObjectMeta.Namespace
-	ownerRef := generateOwnerRef(api)
+	ownerRef := GenerateOwnerRef(api)
 
-	arBuilder := builders.RequestAuthenticationBuilder().
+	arBuilder := builders.AuthorizationPolicyBuilder().
 		GenerateName(namePrefix).
 		Namespace(namespace).
 		Owner(builders.OwnerReference().From(&ownerRef)).
-		Spec(builders.RequestAuthenticationSpecBuilder().From(generateRequestAuthenticationSpec(api, rule))).
+		Spec(builders.AuthorizationPolicySpecBuilder().From(generateAuthorizationPolicySpec(api, rule))).
 		Label(OwnerLabel, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace)).
 		Label(OwnerLabelv1alpha1, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace))
 
@@ -33,8 +34,7 @@ func generateRequestAuthentication(api *gatewayv1beta1.APIRule, rule gatewayv1be
 	return arBuilder.Get()
 }
 
-func generateRequestAuthenticationSpec(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule) *v1beta1.RequestAuthentication {
-
+func generateAuthorizationPolicySpec(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule) *v1beta1.AuthorizationPolicy {
 	var serviceName string
 	if rule.Service != nil {
 		serviceName = *rule.Service.Name
@@ -42,9 +42,12 @@ func generateRequestAuthenticationSpec(api *gatewayv1beta1.APIRule, rule gateway
 		serviceName = *api.Spec.Service.Name
 	}
 
-	requestAuthenticationSpec := builders.RequestAuthenticationSpecBuilder().
+	authorizationPolicySpec := builders.AuthorizationPolicySpecBuilder().
 		Selector(builders.SelectorBuilder().MatchLabels("app", serviceName)).
-		JwtRules(builders.JwtRuleBuilder().From(rule.AccessStrategies))
+		Rule(builders.RuleBuilder().
+			RuleFrom(builders.RuleFromBuilder().Source()).
+			RuleTo(builders.RuleToBuilder().
+				Operation(builders.OperationBuilder().Methods(rule.Methods).Path(rule.Path))))
 
-	return requestAuthenticationSpec.Get()
+	return authorizationPolicySpec.Get()
 }
